Treemap migration done. Work also done with stacked line. There was no case to test it, though. Moving on and bugfixing when used.
